untangle some unecessarily complicated import chains
