+++ virtualizarr/parsers/kerchunk/translator.py
@@ -4,10 +4,10 @@ from collections.abc import Iterable
 from typing import cast
 
 import numpy as np
-from numcodecs.abc import Codec
 from zarr.core.common import JSON
 from zarr.core.metadata import ArrayV3Metadata
-from zarr.core.metadata.v2 import ArrayV2Metadata
+import ujson
+
 
 from virtualizarr.codecs import (
     numcodec_config_to_configurable,
@@ -24,26 +24,6 @@ from virtualizarr.types.kerchunk import (
     KerchunkStoreRefs,
 )
 from virtualizarr.utils import determine_chunk_grid_shape
-
-
-def to_kerchunk_json(v2_metadata: ArrayV2Metadata) -> str:
-    """Convert V2 metadata to kerchunk JSON format."""
-    import json
-
-    from virtualizarr.writers.kerchunk import NumpyEncoder
-
-    zarray_dict: dict[str, JSON] = v2_metadata.to_dict()
-    if v2_metadata.filters:
-        zarray_dict["filters"] = [
-            # we could also cast to json, but get_config is intended for serialization
-            codec.get_config()
-            for codec in v2_metadata.filters
-            if codec is not None
-        ]  # type: ignore[assignment]
-    if isinstance(compressor := v2_metadata.compressor, Codec):
-        zarray_dict["compressor"] = compressor.get_config()
-
-    return json.dumps(zarray_dict, separators=(",", ":"), cls=NumpyEncoder)
 
 
 def from_kerchunk_refs(decoded_arr_refs_zarray, zattrs) -> "ArrayV3Metadata":
@@ -317,7 +297,6 @@ def fully_decode_arr_refs(d: dict) -> KerchunkArrRefs:
     """
     Only have to do this because kerchunk.SingleHdf5ToZarr apparently doesn't bother converting .zarray and .zattrs contents to dicts, see https://github.com/fsspec/kerchunk/issues/415 .
     """
-    import ujson
 
     sanitized = d.copy()
     for k, v in d.items():
